Arm backend: Move rescale ops out of comparison visitors
Some TOSA ops do not support INT8 as inputs and outputs. Instead, only
INT32 is supported as a whole number type. Prior to this patch, affected
node visitors inserted rescale ops between the data types INT8 and INT32
before and after the operator such that it will accept its input and
output.

Change this by moving the insertion of the rescale ops to a new pass
called InsertRescaleInt32Pass. This will further enable optimizations to
the graph by fusing the rescale nodes.

Only comparison operators are handled in this patch; the remaining ones
are left out to be done in another patch.

Signed-off-by: Martin Lindström <Martin.Lindstroem@arm.com>
Change-Id: I6bb8a10a0b453ae9fd8b8604d64cc5103a4da050